Moves gambit configs into config/lib
