+++ config/lib/gambit-campaigns.js
