+++ config/lib/gambit-conversations.js
