+++ lib/gambit-campaigns.js
@@ -2,7 +2,7 @@
 
 const superagent = require('superagent');
 const logger = require('heroku-logger');
-const config = require('../config/gambit-campaigns');
+const config = require('../config/lib/gambit-campaigns');
 
 /**
  * @param {string} path
