+++ lib/gambit-conversations.js
@@ -2,7 +2,7 @@
 
 const superagent = require('superagent');
 const logger = require('heroku-logger');
-const config = require('../config/gambit-conversations');
+const config = require('../config/lib/gambit-conversations');
 
 /**
  * @param {object} res - Express response
